requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
click==7.0
jinja2==2.10.1
markupsafe==1.1.1         # via jinja2
pyyaml==5.1.1
